requirements.txt
# Core Data Processing
pydantic==2.5.0          # Data validation using Python type hints
pydantic-settings==2.1.0 # Configuration management for Pydantic

# Database Connectivity
mysql-connector-python==8.2.0  # MySQL client library for Python
sqlalchemy==2.0.23             # SQL toolkit and ORM for database abstraction

# ETL & Data Processing
pandas==2.1.3            # Data manipulation and analysis
python-dotenv==1.0.0     # Load environment variables from .env files

# Logging & Utilities
python-json-logger==2.0.7  # JSON logging for structured logs
word2number==1.1           # Convert English text numbers (e.g., "Five") to digits (5)

